rename: animation -> lottie animation 원본 리소스 출처 명시
