+++ public/lottie-animation/graphic.json
